requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-Testing==0.4.1
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.6
Werkzeug==0.9.6
gunicorn==19.0.0
httplib2==0.9
itsdangerous==0.24
nose==1.3.3
psycopg2==2.5.3
python-instagram==1.1.2
simplejson==3.5.3
wsgiref==0.1.2
